added functions to test dates; more tolerant xsd schema; updated TODOs
