+++ src/app/js/lib/utils.js
@@ -11,6 +11,108 @@ exports.getWebPage = function (url, callback) {
     });
 }
 
+exports.verificaPartitaIva = function (piva) {
+    piva = piva.trim();
+    if (piva.length !== 11) return false;
+    if (piva.match(/^\d+$/) === null) return false;
+    let arr = Array.from(piva, x => parseInt(x));
+    let sum = 0;
+    for (let i = 0; i < 11; i++) {
+        if (i % 2 === 0) {
+            sum += arr[i];
+        } else {
+            sum = arr[i]*2  >= 10 ? sum + arr[i]*2 - 9 : sum + arr[i]*2;
+        }
+    }
+    if (sum % 10 === 0) return true;
+    return false;
+}
+
+exports.verificaCodiceFiscale = function (cf) {
+    cf = cf.trim();
+    cf = cf.toUpperCase();
+    if (cf.length !== 16) return false;
+    // if (cf.match(/^\d+$/) === null) return false;
+    let arr = Array.from(cf, x => isNaN(x) ? x : parseInt(x));
+    let sum = 0;
+    let dic = {
+        '0' : {p: 0,d: 1},
+        '1' : {p: 1,d: 0},
+        '2' : {p: 2,d: 5},
+        '3' : {p: 3,d: 7},
+        '4' : {p: 4,d: 9},
+        '5' : {p: 5,d: 13},
+        '6' : {p: 6,d: 15},
+        '7' : {p: 7,d: 17},
+        '8' : {p: 8,d: 19},
+        '9' : {p: 9,d: 21},
+        'A' : {p: 0,d: 1},
+        'B' : {p: 1,d: 0},
+        'C' : {p: 2,d: 5},
+        'D' : {p: 3,d: 7},
+        'E' : {p: 4,d: 9},
+        'F' : {p: 5,d: 13},
+        'G' : {p: 6,d: 15},
+        'H' : {p: 7,d: 17},
+        'I' : {p: 8,d: 19},
+        'J' : {p: 9,d: 21},
+        'K' : {p: 10,d: 2},
+        'L' : {p: 11,d: 4},
+        'M' : {p: 12,d: 18},
+        'N' : {p: 13,d: 20},
+        'O' : {p: 14,d: 11},
+        'P' : {p: 15,d: 3},
+        'Q' : {p: 16,d: 6},
+        'R' : {p: 17,d: 8},
+        'S' : {p: 18,d: 12},
+        'T' : {p: 19,d: 14},
+        'U' : {p: 20,d: 16},
+        'V' : {p: 21,d: 10},
+        'W' : {p: 22,d: 22},
+        'X' : {p: 23,d: 25},
+        'Y' : {p: 24,d: 24},
+        'Z' : {p: 25,d: 23},
+    }
+    let controllo = {
+        '0' : 'A',
+        '1' : 'B',
+        '2' : 'C',
+        '3' : 'D',
+        '4' : 'E',
+        '5' : 'F',
+        '6' : 'G',
+        '7' : 'H',
+        '8' : 'I',
+        '9' : 'J',
+        '10' : 'K',
+        '11' : 'L',
+        '12' : 'M',
+        '13' : 'N',
+        '14' : 'O',
+        '15' : 'P',
+        '16' : 'Q',
+        '17' : 'R',
+        '18' : 'S',
+        '19' : 'T',
+        '20' : 'U',
+        '21' : 'V',
+        '22' : 'W',
+        '23' : 'X',
+        '24' : 'Y',
+        '25' : 'Z'
+    }
+    for (i = 1; i < 16; i ++){
+        if (i % 2 === 0) {
+            sum += dic[arr[i - 1]].p // converto i pari
+        } else {
+            sum += dic[arr[i - 1]].d
+        }
+    }
+    let resto = sum % 26;
+    if (controllo[resto] === arr[15]) return true;
+    return false;
+}
+
 
 smartCigValidity = function (string) {
     var checkStr = string.toUpperCase().substring(1,3);
